Backend/requirements.txt
flask
pandas
numpy
psycopg2-binary
python-dotenv